+++ gradle-plugin/src/functionalTest/kotlin/dev/zacsweers/metro/gradle/incremental/MetroOptionOverrides.kt
@@ -1,0 +1,5 @@
+package dev.zacsweers.metro.gradle.incremental
+
+data class MetroOptionOverrides(
+  val enableInjectConstructorHints: Boolean? = null,
+)